[TS] Part2: SweepQueue writers impl (#3104)
* Added necessary data types

* Added SerDeser test

* Small refactor

* Unifnished version of KvsSweepQueueWriter, very much still WiP

* Addressed comments

* Added most of the functionality, still some refactoring to do

* Finished implementing writers

* Fixed compilation, broken InMemory tests

* Reverted unnecessary gradle import

* Added tests for WriteInfoPartitioner

* Improve tests

* Fix nit

* Simlified constructor of KvsSweepQueuePersister and added a factory method

* Addressed comments

+++ atlasdb-api/src/main/java/com/palantir/atlasdb/keyvalue/api/TargetedSweepMetadata.java
@@ -26,7 +26,7 @@ public abstract class TargetedSweepMetadata implements Persistable {
     public abstract boolean conservative();
     public abstract boolean dedicatedRow();
     public abstract int shard();
-    public abstract int dedicatedRowNumber();
+    public abstract long dedicatedRowNumber();
 
     @Value.Check
     void checkShardSize() {
